implementación de funcionalidad
Se movió el modulo de registro a el panel administrador para que se de el control total de registrar a los usuarios.
