+++ app/templates/Dashboard/register.html
@@ -2,6 +2,9 @@
 
 {% load static %}
 {% block login %}
+<div class="container-fluid">
+    <a href="{% url 'Usuarios' %}" class="m-auto" style="text-decoration: none; color: black;">❰</a>
+</div>
 <h2 class="text-center mb-4" style="font-family: 'Lucida Sans';">Regístrate</h2>
 <form method="POST">
     {% csrf_token %}
@@ -35,11 +38,7 @@
         <input required name="password" type="password" class="form-control border border-secondary">
     </div>
     <div class="mt-4 d-grid">
-        <button class="botoncitos btn btn-primary" style="background-color: #39a900; border-color: transparent;" type="submit">Registrate</button>
+        <button class="botoncitos btn btn-primary" style= "border-color: transparent;" type="submit">Registrar</button>
     </div>
 </form>
-<div class="mt-3">
-    <p class="mb-0  text-center">Ya tienes una cuenta? <a href="{% url 'login' %}"
-            class="textito text-primary fw-bold" style="text-decoration: none;">Ingresar</a></p>
-</div>
 {% endblock %}
