+++ app/templates/login/login.html
@@ -36,10 +36,6 @@
                     </div>
                     <input type="hidden" name="newpassword" value="True">
                 </form>
-                <div class="mt-3">
-                    <p class="mb-0  text-center">No tienes una cuenta? <a href="{% url 'register' %}"
-                            class="text-primary fw-bold" style="text-decoration: none;">Registrate</a></p>
-                </div>
                 {% endblock %}
             </div>
         </div>
