+++ project/urls.py
@@ -19,6 +19,7 @@ from django.urls import path
 from app import views
 from django.conf.urls.static import static
 from django.conf import settings
+from app.views.dashboard import eliminar_usuario
 
 
 
@@ -85,4 +86,5 @@ urlpatterns = [
     path('Dashboard/Proyectos-Inactivos/', views.proyectosINA, name='Proyecto-inactivo'),
     path('Dashboard/Proyectos-Completos/', views.proyectoT, name='Proyecto-Completo'),
     path('Dashboard/Proyectos-Pendientes', views.proyectoP, name='Proyecto-Pendiente'),
+    path('eliminar_usuario/<int:usuario_id>/', eliminar_usuario, name='eliminar_usuario'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
